requirements.txt
streamlit
google-generativeai
openai-whisper
pymilvus
PyPDF2
crewai
crewai-tools